CLI: Consolidate sign-only helpers
