+++ cli/src/cli.rs
@@ -1,10 +1,10 @@
 use crate::{
     checks::*,
-    cli_output::{CliAccount, CliSignOnlyData, CliSignature, OutputFormat},
+    cli_output::{CliAccount, CliSignature, OutputFormat},
     cluster_query::*,
     display::{new_spinner_progress_bar, println_name_value, println_transaction},
     nonce::*,
-    offline::blockhash_query::BlockhashQuery,
+    offline::{blockhash_query::BlockhashQuery, return_signers},
     spend_utils::*,
     stake::*,
     validator_info::*,
@@ -907,35 +907,6 @@ pub fn get_blockhash_and_fee_calculator(
     } else {
         rpc_client.get_recent_blockhash()?
     })
-}
-
-pub fn return_signers(tx: &Transaction, output_format: &OutputFormat) -> ProcessResult {
-    let verify_results = tx.verify_with_results();
-    let mut signers = Vec::new();
-    let mut absent = Vec::new();
-    let mut bad_sig = Vec::new();
-    tx.signatures
-        .iter()
-        .zip(tx.message.account_keys.iter())
-        .zip(verify_results.into_iter())
-        .for_each(|((sig, key), res)| {
-            if res {
-                signers.push(format!("{}={}", key, sig))
-            } else if *sig == Signature::default() {
-                absent.push(key.to_string());
-            } else {
-                bad_sig.push(key.to_string());
-            }
-        });
-
-    let cli_command = CliSignOnlyData {
-        blockhash: tx.message.recent_blockhash.to_string(),
-        signers,
-        absent,
-        bad_sig,
-    };
-
-    Ok(output_format.formatted_string(&cli_command))
 }
 
 pub fn parse_create_address_with_seed(
@@ -2351,9 +2322,7 @@ mod tests {
     use solana_client::mock_sender::SIGNATURE;
     use solana_sdk::{
         pubkey::Pubkey,
-        signature::{
-            keypair_from_seed, read_keypair_file, write_keypair_file, NullSigner, Presigner,
-        },
+        signature::{keypair_from_seed, read_keypair_file, write_keypair_file, Presigner},
         transaction::TransactionError,
     };
     use std::path::PathBuf;
@@ -3132,53 +3101,5 @@ mod tests {
                 ],
             }
         );
-    }
-
-    #[test]
-    fn test_return_signers() {
-        struct BadSigner {
-            pubkey: Pubkey,
-        }
-
-        impl BadSigner {
-            pub fn new(pubkey: Pubkey) -> Self {
-                Self { pubkey }
-            }
-        }
-
-        impl Signer for BadSigner {
-            fn try_pubkey(&self) -> Result<Pubkey, SignerError> {
-                Ok(self.pubkey)
-            }
-
-            fn try_sign_message(&self, _message: &[u8]) -> Result<Signature, SignerError> {
-                Ok(Signature::new(&[1u8; 64]))
-            }
-        }
-
-        let present: Box<dyn Signer> = Box::new(keypair_from_seed(&[2u8; 32]).unwrap());
-        let absent: Box<dyn Signer> = Box::new(NullSigner::new(&Pubkey::new(&[3u8; 32])));
-        let bad: Box<dyn Signer> = Box::new(BadSigner::new(Pubkey::new(&[4u8; 32])));
-        let to = Pubkey::new(&[5u8; 32]);
-        let nonce = Pubkey::new(&[6u8; 32]);
-        let from = present.pubkey();
-        let fee_payer = absent.pubkey();
-        let nonce_auth = bad.pubkey();
-        let mut tx = Transaction::new_unsigned(Message::new_with_nonce(
-            vec![system_instruction::transfer(&from, &to, 42)],
-            Some(&fee_payer),
-            &nonce,
-            &nonce_auth,
-        ));
-
-        let signers = vec![present.as_ref(), absent.as_ref(), bad.as_ref()];
-        let blockhash = Hash::new(&[7u8; 32]);
-        tx.try_partial_sign(&signers, blockhash).unwrap();
-        let res = return_signers(&tx, &OutputFormat::JsonCompact).unwrap();
-        let sign_only = parse_sign_only_reply_string(&res);
-        assert_eq!(sign_only.blockhash, blockhash);
-        assert_eq!(sign_only.present_signers[0].0, present.pubkey());
-        assert_eq!(sign_only.absent_signers[0], absent.pubkey());
-        assert_eq!(sign_only.bad_signers[0], bad.pubkey());
     }
 }
